LayoutTests/platform/glib/imported/w3c/web-platform-tests/webcodecs/transfering.https.any-expected.txt


FAIL Test transfering ArrayBuffer to VideoFrame assert_equals: data.length after detach expected 0 but got 16
PASS Test transfering detached buffer to VideoFrame
FAIL Test transfering view of an ArrayBuffer to VideoFrame assert_equals: data.length after detach expected 0 but got 16
PASS Test transfering same array buffer twice
FAIL Test transfering ArrayBuffer to EncodedAudioChunk assert_equals: data.length after detach expected 0 but got 10
FAIL Test transfering ArrayBuffer to EncodedVideoChunk assert_equals: data.length after detach expected 0 but got 10
FAIL Test transfering ArrayBuffer to AudioData assert_equals: data.length after detach expected 0 but got 10
FAIL Encoding from AudioData with transferred buffer assert_equals: buffer.length after detach expected 0 but got 9600
FAIL Test transfering ArrayBuffer to ImageDecoder. promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: ImageDecoder"
